refactor: extract format-string date detection out of DateUtils
